backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
supabase==2.7.4
groq
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.7.4
pydantic-settings==2.3.4
openai
PyJWT
